refactor: Move AmaiSosu.Compilation.Compiler and AmaiSosu.Installation.Installer classes to AmaiSosu.Core namespace refactor: Rename PackageFactory's Installation Path parameters to 'path' refactor: remove unnecessary Using from AmaiSosu.Startup
