+++ src/AmaiSosu/AmaiSosu.csproj
@@ -48,13 +48,13 @@
     <Compile Include="Common\Package.cs" />
     <Compile Include="Common\Paths.cs" />
     <Compile Include="Common\Verification.cs" />
-    <Compile Include="Compilation\Compiler.cs" />
+    <Compile Include="Core\Compiler.cs" />
     <Compile Include="Context.cs" />
     <Compile Include="Detection\Executable.cs" />
     <Compile Include="Detection\ExecutableFactory.cs" />
     <Compile Include="Detection\Loader.cs" />
     <Compile Include="Detection\LoaderException.cs" />
-    <Compile Include="Installation\Installer.cs" />
+    <Compile Include="Core\Installer.cs" />
     <Compile Include="Common\PackageFactory.cs" />
     <Compile Include="Common\IO\Copy.cs" />
     <Compile Include="Common\IO\Move.cs" />
